PC1/mqtt_to_mongo/requirements.txt
paho-mqtt==2.1.0 
pymongo==4.10.1 
pydantic==2.10.6
pytz==2025.1
typing==3.7.4.3
python-dateutil
